pyproject.toml
[project]
name = "discord-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "discord-py>=2.5.2",
    "injector>=0.22.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
]

[tool.coverage.run]
source = ["discord-app", "tests"]
